- Collapse indexed only into streaming and account for difference there.
